requirements.txt
PlexAPI==4.10.1
requests==2.26.0
websocket-client==1.3.2
PyYAML==6.0.1
